Add titles in top movies dialog
