+++ lib/l10n/app_en.arb
@@ -19,10 +19,13 @@
     "delete_all_searchers": "Delete all searchers",
     "title_bookmarks_page": "Your bookmarks",
     "title_top_movies_page": "If you don't know to see",
-    "exclude_animation": "Exclude animation",
-    "order_by": "Order by",
-    "delete_filters": "Delete filters",
-    "apply_filters": "Apply filters",
+    "title_platforms_dialog": "What platforms do you have?",
+    "title_years_dialog": "What years do you want to see?",
+    "title_exclude_animation_dialog": "Exclude animation",
+    "title_order_by_dialog": "Order by",
+    "title_genres_dialog": "What genres does it have to include?",
+    "delete_filters_dialog": "Delete filters",
+    "apply_filters_dialog": "Apply filters",
     "add_bookmark": "Movie added successfully",
     "delete_bookmark": "Movie removed successfully",
     "timeout_error": "Ups! Something went wrong with the server.",
