+++ lib/l10n/app_es.arb
@@ -19,10 +19,13 @@
   "delete_all_searchers": "Eliminar búsquedas recientes",
   "title_bookmarks_page": "Tus marcadores",
   "title_top_movies_page": "Si no sabes qué ver",
-  "exclude_animation": "Excluir animación",
-  "order_by": "Ordenar por",
-  "delete_filters": "Eliminar filtros",
-  "apply_filters": "Aplicar filtros",
+  "title_platforms_dialog": "¿Qué plataformas tienes?",
+  "title_years_dialog": "¿Qué años te interesan?",
+  "title_exclude_animation_dialog": "Excluir animación",
+  "title_order_by_dialog": "Ordenar por",
+  "title_genres_dialog": "¿Qué géneros tiene que incluir?",
+  "delete_filters_dialog": "Eliminar filtros",
+  "apply_filters_dialog": "Aplicar filtros",
   "add_bookmark": "Película añadida correctamente",
   "delete_bookmark": "Película eliminada correctamente",
   "timeout_error": "¡Ups! Parece que ha habido un problema con el servidor.",
