+++ lib/utils/justwatch.dart
@@ -1,5 +1,6 @@
 class JustwatchAssets {
   static final Map<String, List<String>> justwatchAssets = {
+    'acontra': ['acontra plus'],
     'amazon': ['amazon video', 'amazon prime video'],
     'apple': ['apple itunes'],
     'disney': ['disney plus'],
